+++ src/utils/index.ts
@@ -3,4 +3,6 @@ export { getTreeItem } from "./getTreeItem";
 export { isAncestor } from "./isAncestor";
 export { isDroppable } from "./isDroppable";
 export { mutateTree } from "./mutateTree";
+export { mutateTreeWithIndex } from "./mutateTreeWithIndex";
 export { getDropTarget } from "./getDropTarget";
+export { getModifiedIndex } from "./getModifiedIndex";