+++ src/utils/mutateTreeWithIndex.ts
@@ -1,0 +1,51 @@
+import { produce } from "immer";
+import { NodeModel } from "../types";
+import { getModifiedIndex } from "./getModifiedIndex";
+
+const arrayMoveMutable = <T>(
+  array: T[],
+  fromIndex: number,
+  toIndex: number
+) => {
+  const startIndex = fromIndex < 0 ? array.length + fromIndex : fromIndex;
+
+  if (startIndex >= 0 && startIndex < array.length) {
+    const endIndex = toIndex < 0 ? array.length + toIndex : toIndex;
+
+    const [item] = array.splice(fromIndex, 1);
+    array.splice(endIndex, 0, item);
+  }
+};
+
+const arrayMoveImmutable = <T>(
+  array: T[],
+  fromIndex: number,
+  toIndex: number
+) => {
+  array = [...array];
+  arrayMoveMutable(array, fromIndex, toIndex);
+  return array;
+};
+
+export const mutateTreeWithIndex = <T>(
+  tree: NodeModel<T>[],
+  dragSourceId: NodeModel["id"],
+  dropTargetId: NodeModel["id"],
+  index: number
+): NodeModel<T>[] => {
+  const [srcIndex, destIndex] = getModifiedIndex(
+    tree,
+    dragSourceId,
+    dropTargetId,
+    index
+  );
+  const newTree = arrayMoveImmutable(tree, srcIndex, destIndex);
+
+  return produce(newTree, (draft) => {
+    draft.forEach((node) => {
+      if (node.id === dragSourceId) {
+        node.parent = dropTargetId;
+      }
+    });
+  });
+};